functionnal free and allocation

+++ ft_malloc.h
@@ -6,7 +6,7 @@
 /*   By: kcabus <kcabus@student.le-101.fr>          +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2019/08/08 13:35:53 by kcabus       #+#   ##    ##    #+#       */
-/*   Updated: 2020/02/06 12:23:29 by kcabus      ###    #+. /#+    ###.fr     */
+/*   Updated: 2020/02/06 14:38:25 by kcabus      ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
@@ -73,7 +73,7 @@ void				*large_alloc(t_page *page, size_t size);
 void				ft_free(void *ptr);
 
 size_t				get_size_align(size_t size);
-int					get_size_alloc(int type);
+size_t				get_size_alloc(int type);
 size_t				get_size_page(int type);
 int					page_is_empty(t_page *page);
